Split calculator and timer into separate files
